pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-gateway"
version = "0.1.0"
description = "MCP Gateway - aggregate multiple MCP servers behind a single Streamable HTTP endpoint"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "mcp>=1.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "click>=8.0",
    "anyio>=4.0",
]

[project.scripts]
mcp-gateway = "mcp_gateway.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gateway"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
